+++ m2r2/__init__.py
@@ -1,0 +1,3 @@
+from .m2r2 import M2R, convert
+
+__all__ = ("M2R", "convert")
