+++ m2r2/constants.py
@@ -1,0 +1,5 @@
+PROLOG = """\
+.. role:: raw-html-m2r(raw)
+   :format: html
+
+"""
