+++ m2r2/rst/directives.py
@@ -1,0 +1,92 @@
+import os
+
+from docutils import io, nodes, statemachine, utils
+from docutils.parsers import rst
+
+from m2r2 import M2R
+
+
+class MdInclude(rst.Directive):
+    """Directive class to include markdown in sphinx.
+
+    Load a file and convert it to rst and insert as a node. Currently
+    directive-specific options are not implemented.
+    """
+
+    required_arguments = 1
+    optional_arguments = 0
+    option_spec = {
+        "start-line": int,
+        "end-line": int,
+    }
+
+    def run(self):
+        """Most of this method is from ``docutils.parser.rst.Directive``.
+
+        docutils version: 0.12
+        """
+        if not self.state.document.settings.file_insertion_enabled:
+            raise self.warning('"%s" directive disabled.' % self.name)
+        source = self.state_machine.input_lines.source(
+            self.lineno - self.state_machine.input_offset - 1
+        )
+        source_dir = os.path.dirname(os.path.abspath(source))
+        path = rst.directives.path(self.arguments[0])
+        path = os.path.normpath(os.path.join(source_dir, path))
+        path = utils.relative_path(None, path)
+        path = nodes.reprunicode(path)
+
+        # get options (currently not use directive-specific options)
+        encoding = self.options.get(
+            "encoding", self.state.document.settings.input_encoding
+        )
+        e_handler = self.state.document.settings.input_encoding_error_handler
+        tab_width = self.options.get(
+            "tab-width", self.state.document.settings.tab_width
+        )
+
+        # open the including file
+        try:
+            self.state.document.settings.record_dependencies.add(path)
+            include_file = io.FileInput(
+                source_path=path, encoding=encoding, error_handler=e_handler
+            )
+        except UnicodeEncodeError:
+            raise self.severe(
+                'Problems with "%s" directive path:\n'
+                'Cannot encode input file path "%s" '
+                "(wrong locale?)." % (self.name, str(path))
+            )
+        except IOError as error:
+            raise self.severe(
+                'Problems with "%s" directive path:\n%s.'
+                % (self.name, io.error_string(error))
+            )
+
+        # read from the file
+        startline = self.options.get("start-line", None)
+        endline = self.options.get("end-line", None)
+        try:
+            if startline or (endline is not None):
+                lines = include_file.readlines()
+                rawtext = "".join(lines[startline:endline])
+            else:
+                rawtext = include_file.read()
+        except UnicodeError as error:
+            raise self.severe(
+                'Problem with "%s" directive:\n%s' % (self.name, io.error_string(error))
+            )
+
+        config = self.state.document.settings.env.config
+        converter = M2R(
+            no_underscore_emphasis=config.no_underscore_emphasis,
+            parse_relative_links=config.m2r_parse_relative_links,
+            anonymous_references=config.m2r_anonymous_references,
+            disable_inline_math=config.m2r_disable_inline_math,
+            use_mermaid=config.m2r_use_mermaid,
+        )
+        include_lines = statemachine.string2lines(
+            converter(rawtext), tab_width, convert_whitespace=True
+        )
+        self.state_machine.insert_input(include_lines, path)
+        return []
