+++ m2r2/typing.py
@@ -1,0 +1,5 @@
+from typing import Any, Dict, Tuple
+
+State = Dict[str, Any]
+Token = Dict[str, Any]
+Element = Tuple[str, ...]
