+++ requirements-test.txt
@@ -1,4 +1,3 @@
-mock; python_version < '3.3'
 pytest
 pytest-cov
 six
